% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleVision-LIB.R
\name{checkAndLoadPackages}
\alias{checkAndLoadPackages}
\title{helper function to load required packages}
\usage{
checkAndLoadPackages()
}
\description{
Thanks to http://stackoverflow.com/questions/4090169/elegant-way-to-check-for-missing-packages-and-install-them
}

