% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleVision-LIB.R
\name{extractResponse}
\alias{extractResponse}
\title{helper function code to extract the response data.frame}
\usage{
extractResponse(pp, feature)
}
\arguments{
\item{provide}{the API rest response, and the feature request}
}
\value{
get the data frame back
}
\description{

}

