% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleVision-LIB.R
\name{getGoogleVisionResponse}
\alias{getGoogleVisionResponse}
\title{Calling Google's Cloud Vision API}
\usage{
getGoogleVisionResponse(imagePath, feature = "LABEL_DETECTION",
  numResults = 5)
}
\arguments{
\item{path}{or url to the image}

\item{feature:}{one out of: FACE_DETECTION, LANDMARK_DETECTION, LOGO_DETECTION, LABEL_DETECTION, TEXT_DETECTION}
}
\value{
data frame with results
}
\description{
input an image, provide the feature type and maxNumber of responses
}
\examples{
getGoogleVisionResponse(imagePath="brandlogos.png", feature="LOGO_DETECTION")

}

