% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleVision-LIB.R
\name{imageToText}
\alias{imageToText}
\title{helper function base_encode code the image file}
\usage{
imageToText(imagePath)
}
\arguments{
\item{imagePath}{provide path/url to image}
}
\value{
get the image back as encoded file
}
\description{
base64 encodes an image file
}

